requirements.txt

openai
fastapi
uvicorn
numpy
pydantic
python-dotenv

openai
python-dotenv
numpy
fastapi
uvicorn
pydantic
